refactor(search): co-locate advanced form schema
